backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.6.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
aiofiles==23.2.1
opencv-python==4.9.0.80
pillow==11.0.0
numpy==1.26.3
ultralytics==8.1.0
requests==2.31.0
fastapi-cors==0.0.6
setuptools==69.0.0



